Incoming Fern-generated changes inlined websocket logic and removed
websocket_url param from connect(). Kept our version which uses the
open_*() helpers with reconnect support. Non-conflicting changes
(dry_run param, config_options, new types) preserved from incoming.

+++ src/phonic/requests/tool_call_output_processed_payload.py
@@ -30,7 +30,7 @@ class ToolCallOutputProcessedPayloadParams(typing_extensions.TypedDict):
     Webhook endpoint URL (null for WebSocket tools)
     """
 
-    endpoint_timeout_ms: typing_extensions.NotRequired[typing.Optional[float]]
+    endpoint_timeout_ms: typing_extensions.NotRequired[typing.Optional[int]]
     """
     Webhook timeout in milliseconds (null for WebSocket tools)
     """
@@ -65,7 +65,7 @@ class ToolCallOutputProcessedPayloadParams(typing_extensions.TypedDict):
     WebSocket tool output (null for webhook tools)
     """
 
-    response_status_code: typing_extensions.NotRequired[typing.Optional[float]]
+    response_status_code: typing_extensions.NotRequired[typing.Optional[int]]
     """
     Webhook HTTP status code (null for WebSocket tools)
     """
